requirements.txt
tqdm
torch
accelerate
mlflow
loguru
pyyaml
boto3
tensorboard

-e .